refactor(registry): split registry index into ui/blocks/types modules
Mirror shadcn's split registry-*.ts layout. The monolithic registry-meta.ts
is partitioned by item category while preserving every existing export, so
the showcase imports are unchanged.

- registry-types.ts  — shared RegistryEntryMeta / category / block types.
- registry-ui.ts     — UI_REGISTRY (28 component entries + their demos).
- registry-blocks.ts — BLOCK_REGISTRY (26 block entries + their sources).
- registry-meta.ts   — combiner: re-exports the types, builds
  REGISTRY = [...UI_REGISTRY, ...BLOCK_REGISTRY], and keeps the derived
  helpers (REGISTRY_BY_NAME/CATEGORY, BLOCKS_BY_KIND, label/order maps).

tsc --noEmit, next build, and next lint all pass.
