+++ registry/new-york/registry-blocks.ts
@@ -1,0 +1,449 @@
+import AppShell01 from "@/registry/new-york/blocks/app-shell-01/app-shell-01"
+import AppShell02 from "@/registry/new-york/blocks/app-shell-02/app-shell-02"
+import Blog01 from "@/registry/new-york/blocks/blog-01/blog-01"
+import Contact01 from "@/registry/new-york/blocks/contact-01/contact-01"
+import Cta01 from "@/registry/new-york/blocks/cta-01/cta-01"
+import Cta02 from "@/registry/new-york/blocks/cta-02/cta-02"
+import Dashboard01 from "@/registry/new-york/blocks/dashboard-01/dashboard-01"
+import Faq01 from "@/registry/new-york/blocks/faq-01/faq-01"
+import Faq02 from "@/registry/new-york/blocks/faq-02/faq-02"
+import Feature01 from "@/registry/new-york/blocks/feature-01/feature-01"
+import Feature02 from "@/registry/new-york/blocks/feature-02/feature-02"
+import Footer01 from "@/registry/new-york/blocks/footer-01/footer-01"
+import Header01 from "@/registry/new-york/blocks/header-01/header-01"
+import Hero01 from "@/registry/new-york/blocks/hero-01/hero-01"
+import Hero02 from "@/registry/new-york/blocks/hero-02/hero-02"
+import Hero03 from "@/registry/new-york/blocks/hero-03/hero-03"
+import ImageGallery01 from "@/registry/new-york/blocks/image-gallery-01/image-gallery-01"
+import Integrations01 from "@/registry/new-york/blocks/integrations-01/integrations-01"
+import LogoCloud01 from "@/registry/new-york/blocks/logo-cloud-01/logo-cloud-01"
+import Login01 from "@/registry/new-york/blocks/login-01/login-01"
+import Login02 from "@/registry/new-york/blocks/login-02/login-02"
+import NotFound01 from "@/registry/new-york/blocks/not-found-01/not-found-01"
+import Pricing01 from "@/registry/new-york/blocks/pricing-01/pricing-01"
+import Pricing02 from "@/registry/new-york/blocks/pricing-02/pricing-02"
+import Testimonial01 from "@/registry/new-york/blocks/testimonial-01/testimonial-01"
+import Testimonial02 from "@/registry/new-york/blocks/testimonial-02/testimonial-02"
+
+import type { RegistryEntryMeta } from "@/registry/new-york/registry-types"
+
+export const BLOCK_REGISTRY: RegistryEntryMeta[] = [
+  {
+    name: "hero-01",
+    title: "Hero · Split with install card",
+    description:
+      "Split hero with eyebrow tag, display headline, dual CTA, three-stat strip and a mock install-card visual.",
+    blockTagline: "Split layout · stat strip · mock install card",
+    category: "blocks",
+    blockKind: "hero",
+    status: "stable",
+    Demo: Hero01,
+    sourceFiles: ["registry/new-york/blocks/hero-01/hero-01.tsx"],
+    installTargets: ["components/blocks/hero-01.tsx"],
+    registryDependencies: ["button"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "hero-02",
+    title: "Hero · Centered editorial",
+    description:
+      "Centered hero with animated live-pill, display headline with underlined accent, sub-copy and a trusted-by wordmark strip.",
+    blockTagline: "Centered · live pill · wordmark strip",
+    category: "blocks",
+    blockKind: "hero",
+    status: "stable",
+    Demo: Hero02,
+    sourceFiles: ["registry/new-york/blocks/hero-02/hero-02.tsx"],
+    installTargets: ["components/blocks/hero-02.tsx"],
+    registryDependencies: ["button"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "hero-03",
+    title: "Hero · Email capture",
+    description:
+      "Centered hero with a rating pill, display headline, sub-copy, an inline email-capture form with a success state, a feature checklist and an avatar social-proof row.",
+    blockTagline: "Centered · email capture · social proof",
+    category: "blocks",
+    blockKind: "hero",
+    status: "stable",
+    Demo: Hero03,
+    sourceFiles: ["registry/new-york/blocks/hero-03/hero-03.tsx"],
+    installTargets: ["components/blocks/hero-03.tsx"],
+    registryDependencies: ["button", "input"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "feature-01",
+    title: "Feature · Alternating rows",
+    description:
+      "Three alternating feature rows, each pairing a stylized Tailwind-only mock UI with a copy column (eyebrow, headline, paragraph, 3-item checklist).",
+    blockTagline: "Alternating rows · mock UIs · checklist",
+    category: "blocks",
+    blockKind: "feature",
+    status: "stable",
+    Demo: Feature01,
+    sourceFiles: ["registry/new-york/blocks/feature-01/feature-01.tsx"],
+    installTargets: ["components/blocks/feature-01.tsx"],
+    registryDependencies: ["button"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "feature-02",
+    title: "Feature · Icon grid",
+    description:
+      "Section header above a 3-column, 2-row grid of bordered feature cards with lucide icons, headlines and short blurbs.",
+    blockTagline: "Icon grid · 6 cards · concise blurbs",
+    category: "blocks",
+    blockKind: "feature",
+    status: "stable",
+    Demo: Feature02,
+    sourceFiles: ["registry/new-york/blocks/feature-02/feature-02.tsx"],
+    installTargets: ["components/blocks/feature-02.tsx"],
+    registryDependencies: [],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "pricing-01",
+    title: "Pricing · Three tiers",
+    description:
+      "Three-tier card row with a featured middle plan. Each card lists price, blurb, feature checklist and a primary or outline CTA.",
+    blockTagline: "3 tiers · featured plan · checklist",
+    category: "blocks",
+    blockKind: "pricing",
+    status: "stable",
+    Demo: Pricing01,
+    sourceFiles: ["registry/new-york/blocks/pricing-01/pricing-01.tsx"],
+    installTargets: ["components/blocks/pricing-01.tsx"],
+    registryDependencies: ["button"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "pricing-02",
+    title: "Pricing · Comparison table",
+    description:
+      "Compare-by-feature pricing table with sticky tier header (tier name, price, CTA) and ~8 feature rows below.",
+    blockTagline: "Comparison table · sticky header · per-tier CTA",
+    category: "blocks",
+    blockKind: "pricing",
+    status: "stable",
+    Demo: Pricing02,
+    sourceFiles: ["registry/new-york/blocks/pricing-02/pricing-02.tsx"],
+    installTargets: ["components/blocks/pricing-02.tsx"],
+    registryDependencies: ["button"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "testimonial-01",
+    title: "Testimonial · Single quote",
+    description:
+      "Centered single quote with stylized open-mark, author block (avatar, name, role, company) and a muted wordmark row below.",
+    blockTagline: "Single quote · author block · wordmark row",
+    category: "blocks",
+    blockKind: "testimonial",
+    status: "stable",
+    Demo: Testimonial01,
+    sourceFiles: ["registry/new-york/blocks/testimonial-01/testimonial-01.tsx"],
+    installTargets: ["components/blocks/testimonial-01.tsx"],
+    registryDependencies: [],
+    dependencies: [],
+  },
+  {
+    name: "testimonial-02",
+    title: "Testimonial · Masonry grid",
+    description:
+      "Masonry quote grid (CSS columns) with ~6 bordered quote cards of varying length and author rows.",
+    blockTagline: "Masonry grid · 6 quotes · varied length",
+    category: "blocks",
+    blockKind: "testimonial",
+    status: "stable",
+    Demo: Testimonial02,
+    sourceFiles: ["registry/new-york/blocks/testimonial-02/testimonial-02.tsx"],
+    installTargets: ["components/blocks/testimonial-02.tsx"],
+    registryDependencies: [],
+    dependencies: [],
+  },
+  {
+    name: "cta-01",
+    title: "CTA · Framed band",
+    description:
+      "Framed CTA card with corner marks, headline + sub-copy on the left, dual buttons stacked on the right.",
+    blockTagline: "Framed · split layout · corner marks",
+    category: "blocks",
+    blockKind: "cta",
+    status: "stable",
+    Demo: Cta01,
+    sourceFiles: ["registry/new-york/blocks/cta-01/cta-01.tsx"],
+    installTargets: ["components/blocks/cta-01.tsx"],
+    registryDependencies: ["button"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "cta-02",
+    title: "CTA · Centered announcement",
+    description:
+      "Full-bleed centered CTA with framing top/bottom rules, highlight underlay on the key word, and an inline install command.",
+    blockTagline: "Centered · highlight underlay · install command",
+    category: "blocks",
+    blockKind: "cta",
+    status: "stable",
+    Demo: Cta02,
+    sourceFiles: ["registry/new-york/blocks/cta-02/cta-02.tsx"],
+    installTargets: ["components/blocks/cta-02.tsx"],
+    registryDependencies: ["button"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "faq-01",
+    title: "FAQ · Sticky split",
+    description:
+      "Two-column FAQ — sticky heading + contact card on the left, numbered accordion on the right.",
+    blockTagline: "Sticky split · numbered · contact card",
+    category: "blocks",
+    blockKind: "faq",
+    status: "stable",
+    Demo: Faq01,
+    sourceFiles: ["registry/new-york/blocks/faq-01/faq-01.tsx"],
+    installTargets: ["components/blocks/faq-01.tsx"],
+    registryDependencies: ["button", "accordion"],
+    dependencies: ["@radix-ui/react-accordion", "lucide-react"],
+  },
+  {
+    name: "faq-02",
+    title: "FAQ · Centered grid",
+    description:
+      "Centered heading with two-column accordion grid below. Each row tagged with a Qn index.",
+    blockTagline: "Centered · two-column grid · Qn-indexed",
+    category: "blocks",
+    blockKind: "faq",
+    status: "stable",
+    Demo: Faq02,
+    sourceFiles: ["registry/new-york/blocks/faq-02/faq-02.tsx"],
+    installTargets: ["components/blocks/faq-02.tsx"],
+    registryDependencies: ["accordion"],
+    dependencies: ["@radix-ui/react-accordion"],
+  },
+  {
+    name: "login-01",
+    title: "Login · Centered card",
+    description:
+      "Centered login card with monogram, email + password (using the password-input component), remember-me, divider and GitHub / Google providers.",
+    blockTagline: "Centered card · providers · password-input",
+    category: "blocks",
+    blockKind: "login",
+    status: "stable",
+    Demo: Login01,
+    sourceFiles: ["registry/new-york/blocks/login-01/login-01.tsx"],
+    installTargets: ["components/blocks/login-01.tsx"],
+    registryDependencies: ["button", "input", "label", "password-input"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "login-02",
+    title: "Login · Split with testimonial",
+    description:
+      "Two-pane login: form on the left, dark testimonial panel with quote and metrics on the right. Uses the strength-meter variant of password-input.",
+    blockTagline: "Split · testimonial pane · strength meter",
+    category: "blocks",
+    blockKind: "login",
+    status: "stable",
+    Demo: Login02,
+    sourceFiles: ["registry/new-york/blocks/login-02/login-02.tsx"],
+    installTargets: ["components/blocks/login-02.tsx"],
+    registryDependencies: ["button", "input", "label", "password-input"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "header-01",
+    title: "Header · Sticky nav",
+    description:
+      "Sticky top nav with brand monogram, centered anchor links, dual auth CTAs and a slide-down mobile menu.",
+    blockTagline: "Sticky · backdrop blur · mobile menu",
+    category: "blocks",
+    blockKind: "header",
+    status: "stable",
+    Demo: Header01,
+    sourceFiles: ["registry/new-york/blocks/header-01/header-01.tsx"],
+    installTargets: ["components/blocks/header-01.tsx"],
+    registryDependencies: ["button"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "footer-01",
+    title: "Footer · Four columns",
+    description:
+      "Brand + tagline column alongside Product / Company / Resources link columns, with a copyright row and social icons below a thin rule.",
+    blockTagline: "4 columns · social row · copyright",
+    category: "blocks",
+    blockKind: "footer",
+    status: "stable",
+    Demo: Footer01,
+    sourceFiles: ["registry/new-york/blocks/footer-01/footer-01.tsx"],
+    installTargets: ["components/blocks/footer-01.tsx"],
+    registryDependencies: [],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "not-found-01",
+    title: "Not Found · centered 404",
+    description:
+      "Centered 404 with mono eyebrow, display headline, paired CTAs and a 'try one of these' suggested-routes list.",
+    blockTagline: "Centered · paired CTAs · route suggestions",
+    category: "blocks",
+    blockKind: "not-found",
+    status: "stable",
+    Demo: NotFound01,
+    sourceFiles: ["registry/new-york/blocks/not-found-01/not-found-01.tsx"],
+    installTargets: ["components/blocks/not-found-01.tsx"],
+    registryDependencies: ["button"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "logo-cloud-01",
+    title: "Logo Cloud · Bordered grid",
+    description:
+      "Centered eyebrow + headline above a 5-column bordered wordmark grid, with stat strip and case-study link below.",
+    blockTagline: "Bordered grid · 10 wordmarks · stat strip",
+    category: "blocks",
+    blockKind: "logo-cloud",
+    status: "stable",
+    Demo: LogoCloud01,
+    sourceFiles: ["registry/new-york/blocks/logo-cloud-01/logo-cloud-01.tsx"],
+    installTargets: ["components/blocks/logo-cloud-01.tsx"],
+    registryDependencies: ["badge", "button"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "contact-01",
+    title: "Contact · Split form",
+    description:
+      "Production contact form with controlled state, inline validation, character counter, topic select, consent checkbox and a pending/sent state machine. Channel list and remote-location card alongside.",
+    blockTagline: "Validated form · pending · sent state",
+    category: "blocks",
+    blockKind: "contact",
+    status: "stable",
+    Demo: Contact01,
+    sourceFiles: ["registry/new-york/blocks/contact-01/contact-01.tsx"],
+    installTargets: ["components/blocks/contact-01.tsx"],
+    registryDependencies: [
+      "badge",
+      "button",
+      "card",
+      "checkbox",
+      "field",
+      "input",
+      "select",
+      "separator",
+      "textarea",
+    ],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "blog-01",
+    title: "Blog · Featured post and grid",
+    description:
+      "Editorial blog index with a featured post on top and a 4-column grid of post cards underneath. Built on Card and Badge.",
+    blockTagline: "Featured post · 4-up Card grid",
+    category: "blocks",
+    blockKind: "blog",
+    status: "stable",
+    Demo: Blog01,
+    sourceFiles: ["registry/new-york/blocks/blog-01/blog-01.tsx"],
+    installTargets: ["components/blocks/blog-01.tsx"],
+    registryDependencies: ["badge", "button", "card", "separator"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "dashboard-01",
+    title: "Dashboard · Metrics and chart",
+    description:
+      "Operations dashboard with Tabs date-range switcher (1d / 7d / 30d / 90d), 4-up metric strip, weekly bar chart and a recent-activity feed. Data switches live with the range.",
+    blockTagline: "Tabs range · 4 metrics · live data",
+    category: "blocks",
+    blockKind: "dashboard",
+    status: "stable",
+    Demo: Dashboard01,
+    sourceFiles: ["registry/new-york/blocks/dashboard-01/dashboard-01.tsx"],
+    installTargets: ["components/blocks/dashboard-01.tsx"],
+    registryDependencies: ["badge", "button", "card", "separator", "tabs"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "integrations-01",
+    title: "Integrations · Hub and spoke",
+    description:
+      "Two-column integrations section with copy and feature list on the left, orbit diagram (central hub + 7 logo spokes connected by dashed rays) on the right.",
+    blockTagline: "Hub & spoke · 7 spokes · orbit ring",
+    category: "blocks",
+    blockKind: "integrations",
+    status: "stable",
+    Demo: Integrations01,
+    sourceFiles: [
+      "registry/new-york/blocks/integrations-01/integrations-01.tsx",
+    ],
+    installTargets: ["components/blocks/integrations-01.tsx"],
+    registryDependencies: ["badge", "button", "card"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "image-gallery-01",
+    title: "Image Gallery · Filterable masonry",
+    description:
+      "Studio-style masonry gallery with Tabs-driven category filter, real photo tiles via next/image, varied aspect ratios, hover zoom + arrow chip and an EmptyState fallback for empty filters.",
+    blockTagline: "Tabs filter · masonry · empty state",
+    category: "blocks",
+    blockKind: "image-gallery",
+    status: "stable",
+    Demo: ImageGallery01,
+    sourceFiles: [
+      "registry/new-york/blocks/image-gallery-01/image-gallery-01.tsx",
+    ],
+    installTargets: ["components/blocks/image-gallery-01.tsx"],
+    registryDependencies: ["badge", "button", "empty-state", "tabs"],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "app-shell-01",
+    title: "App Shell · Sidebar and topbar",
+    description:
+      "Drop-in admin shell layout built on the shadcn Sidebar primitive: collapsible icon-rail sidebar with nav badges and a footer profile row, sticky topbar with breadcrumb, command-palette search and notification button, plus a live-filtering accounts table in the main area.",
+    blockTagline: "Collapsible Sidebar · sticky topbar · live table",
+    category: "blocks",
+    blockKind: "app-shell",
+    status: "stable",
+    Demo: AppShell01,
+    sourceFiles: ["registry/new-york/blocks/app-shell-01/app-shell-01.tsx"],
+    installTargets: ["components/blocks/app-shell-01.tsx"],
+    registryDependencies: [
+      "badge",
+      "button",
+      "card",
+      "input-group",
+      "kbd",
+      "separator",
+      "sidebar",
+    ],
+    dependencies: ["lucide-react"],
+  },
+  {
+    name: "app-shell-02",
+    title: "App Shell · Top nav and settings",
+    description:
+      "Sidebar-free admin shell with a sticky top navigation bar (logo, primary links, search and avatar) over a settings layout: an in-page vertical nav switches a detail card of definition-list fields with per-field edit actions.",
+    blockTagline: "Top nav · in-page settings nav · detail card",
+    category: "blocks",
+    blockKind: "app-shell",
+    status: "stable",
+    Demo: AppShell02,
+    sourceFiles: ["registry/new-york/blocks/app-shell-02/app-shell-02.tsx"],
+    installTargets: ["components/blocks/app-shell-02.tsx"],
+    registryDependencies: [
+      "badge",
+      "button",
+      "card",
+      "input-group",
+      "separator",
+    ],
+    dependencies: ["lucide-react"],
+  },
+]
